refactor : EntityColumn 추상화, Id 와 일반 컬럼 분리
