+++ src/main/java/persistence/entity/EntityPersister.java
@@ -1,10 +1,7 @@
 package persistence.entity;
 
 import jdbc.JdbcTemplate;
-import persistence.core.EntityColumn;
-import persistence.core.EntityColumns;
-import persistence.core.EntityMetadata;
-import persistence.core.EntityMetadataProvider;
+import persistence.core.*;
 import persistence.sql.dml.DmlGenerator;
 import persistence.util.ReflectionUtils;
 
@@ -12,7 +9,7 @@ import java.util.List;
 
 public class EntityPersister {
     private final String tableName;
-    private final EntityColumn idColumn;
+    private final EntityIdColumn idColumn;
     private final EntityColumns columns;
     private final EntityColumns insertableColumns;
     private final DmlGenerator dmlGenerator;
